+++ protected/controllers/SiteController.php
@@ -352,7 +352,35 @@ class SiteController extends Controller {
 		$this->render('advisory');
 	}
 	public function actionFaq() {
-		$this->render('faq');
+        $model = new ContactForm;
+        // used to scroll the form into view after refresh, in case of server validation errors
+        $hasValidationErrors = false;
+
+        if (isset($_POST['ContactForm'])) {
+            $model->attributes = $_POST['ContactForm'];
+            if ($model->validate()) {
+                try {
+                    Yii::app()->mailService->sendEmail(
+                        Yii::app()->params['adminEmail'],
+                        Yii::app()->params['adminEmail'],
+                        Yii::app()->params['email_prefix'] . "FAQ query: " . $model->subject,
+                        $model->name . " <" . $model->email . ">\n\n" . $model->body
+                    );
+                } catch (Swift_TransportException $ste) {
+                    Yii::log("Problem sending email from FAQ page - " . $ste->getMessage(), "error");
+                }
+                Yii::app()->user->setFlash('submit-question', 'Thank you for contacting us. We will respond to you as soon as possible.');
+                $hasValidationErrors = false;
+                $this->refresh();
+            } else {
+                $hasValidationErrors = true;
+            }
+        }
+
+		$this->render('faq', array(
+            'model' => $model,
+            'hasValidationErrors' => $hasValidationErrors)
+        );
 	}
 
 	public function actionTerm() {
@@ -514,6 +542,4 @@ class SiteController extends Controller {
         echo Yii::app()->newsAndFeedsService->getRss();
         exit;
     }
-
-
 }