+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+- Feat #193: FAQ search and question submission form
+
 ## v4.4.7 - 2025-03-26 - 4c548f037 -
 
 - Fix #2241: Restore Sketchfab visualisations alongside 3D models files visualisations